Details dialog and initial repaint fix
The file details output was previously printed to stdou but now shows up
in a dialog window.

There was an issue with the window painting immediately after opening an
image. A paint event is now manually triggered after file is loaded to
ensure that the image is immediately displayed entirely.